backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
shap>=0.43.0
matplotlib>=3.7.0
seaborn>=0.12.0
ollama>=0.1.0
reportlab>=4.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0
